port some acceptance tests over to copy_example
