+++ testing/example_scripts/conftest_usageerror/conftest.py
@@ -1,0 +1,4 @@
+def pytest_configure(config):
+    import pytest
+
+    raise pytest.UsageError("hello")
