+++ testing/example_scripts/issue88_initial_file_multinodes/conftest.py
@@ -1,0 +1,14 @@
+import pytest
+
+
+class MyFile(pytest.File):
+    def collect(self):
+        return [MyItem("hello", parent=self)]
+
+
+def pytest_collect_file(path, parent):
+    return MyFile(path, parent)
+
+
+class MyItem(pytest.Item):
+    pass
