+++ testing/example_scripts/issue88_initial_file_multinodes/test_hello.py
@@ -1,0 +1,2 @@
+def test_hello():
+    pass
