+++ testing/examples/test_issue519.py
@@ -1,0 +1,3 @@
+def test_510(testdir):
+    testdir.copy_example("issue_519.py")
+    testdir.runpytest("issue_519.py")
